🏗️ 拡張子・ディレクトリ構成・インポート方法の変更・_appを用いた処理共通化
