+++ src/pages/index.jsx
@@ -1,13 +1,11 @@
 import { Inter } from 'next/font/google'
-import { Header } from '@/components/header'
-import Main from '@/components/main'
+import Main from '@/src/components/main/Main'
 
 const inter = Inter({ subsets: ['latin'] })
 
 export default function Home() {
   return (
     <>
-      <Header />
       <Main page="index" />
     </>
   )
